accessibility enhancements
accessibility in webviews is a tricky one. For some reason, whenever
an HTML tag is met, the TTS engine splits. This is a pain because
inflexion points in psalms are indicated with the 'u' (underline) tag.
For instance, in the Psalm 94 of the Laudes, the first verse is read as:

1. Venez crions de 'j'
2. 'o'
3. 'ie' pour le Seigneur

Additionnaly indications like verse numbers and '*' and '+' where read
too which is realy bad for an application whose aim is to help pray and
remove all the clutter.

Here is how we remove it:
- mark verse numbers as aria-hidden=true
- make indications like '*' and '+' as aria-hidden=true
- remove the 'u' tag if a screen reader is detected

All ``aria-hidden=true`` are visible but skipped in screen readers.
Sadly, I could not find an acceptable compromise for the underline
so it's purely and simply stripped away at run time if accessibility
is enabled on the device. Should not be an issue.