get rid of another buggy layer of retry
As an attempt to fight poor connexion quality in some part of the world
I've added a *lot* of retries in every corner of the application. This
brings an unnecessary / rapy complexity with exactly no visible result.

Hence, follow a much more simple approach:
- in case of bad network: fail fast, if DNS fails, there is no reason
anything goes any better
- if Internet connection status changes, we will retry automatically
- the user may retry manually from the menu
- we will add an API to synchronize everything in a single connection,
rather that one per office.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>