fix navigating from bible fragment to lectures fragment
We are adding a Bible section. When navigating from the Bible section to
the Lectures section, the application crashes. This is because we are
calling "LoadLectures" on the new fragment *before* the the fragments
"onCreateView" has been called.

To properly fix this, we use the "arguments" mechanism of the fragment
and re-use as much as possible code from the reload path.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>

fix navigation