[bible] fix flicker with night theme
The night mode was enabled dynamically by togling a class after the page
was loaded. This gave the time to the webview to render with the light
theme at least once so that toggling the dark theme caused the webview
to flicker.

This commit moves to a different approach. Instead, we load a "virtual"
css file. Since the virtual file does not exist in the assets, we can
intercept it in the WebViewClient and dynamically choose the appropriate
CSS file. This also requires to disable the cache. This is OK since all
is local.

The files CSS were batch-patched with:
```
for filename in $(find app/src/main/assets/www -name "*.html")
do
    sed -i 's/assets\/adf2597_dark.front.css/virtual\/application.css/g' "${filename}"
done
```

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>