[core] improve fullscreen toggle UX
We we added the fullscreen mode by default, a major concern was to let
an easy way out of fullscreen mode without cluttering the interface. For
this reason, we added a 'catch all' tap event listener that toggled
fullscreen mode on tap event.

This approaches has a number of drawbacks in a *touch* based interface,
and the fullscreen mode was entered and exited far too frequently. This
is even more visible with the PR on the current verse highlight where
each tap triggers a fullscreen toggle.

This commit is an attempt to fix this while retaining an easily
discoverable exit path.

It completely removes the 'catch all' event listener and insteads binds
the the drawer open/close events. This makes sense since opening the
drawer is part of the 'navigation' intent.

Of course, the usual Android way to diplay the navigation bar by sliding
from the bottom of the screen remains active.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>