[bible] add max width for the offices
On a big tablet screen, the office are not pleasant to read. Most text
is moved far on the left wheras most of the UI is centered.

To make the display nicer on tablets, the width is set to 450dp. The
layout system will then either reduce the size to fit the available
screen, either insert blank spaces on both sides thus centering the
view.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>