[core] hide the navigation bar in landscape mode on tablets
On tablets, the navigation bar in always on the bottom of the screen,
even in ladscape mode. When in landscape mode, we can safely hide to
save some vertical space without breaking the UI.

On all other devices, keep on displaying the bar in landscape mode to
avoid breaking the UI or masking part of the UI when it is displayed in
transluent mode.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>