[core] fix the black share icon
On Android 5.1, the share icon appears black. Technically, we imported a
black share icon and set the "tint" to white et runtime. Since this is
the only occurence of this technique and we always set it to white, this
commit fixes this issue by embedding the white image directly into the
application.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>