add night mode toggle to menu
The night mode is a highly requested feature. It is however hard to
discover. Ideally, we'd add a "FAB" with theme selection and font size
selection, but there is a lot to do before that.

This commit does an intermediate change by adding a checkbox at the top
of the overflow menu. This checkbox allows to toggle between night and
day modes whithout going in the settings menu.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>