Always display the navigation bar
Some user find it hard to exit the application. It is easy for
'smartphone' native people to find the 'swipe' to display the navgation
bar, this is however harder for older people.

Considering the navigation bar is not in itself a source of distraction,
as opposed to the status bar, this commit makes it always visible in
transluent mode, unless the 'full screen' option is enabled.

The content will need to be updated to include a bottom margin so that
the content is not displayed behind the bar.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>