Add margin for the navigation bar at the bottom of webviews
Now that the status bar is set to be displayed all the time, we need to
add some margin at the bottom of the webviews so that the status bar
never overlaps the bottom of the content.

This is done by injecting a bottom margin to the body of the two
webviews.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>