increase webview width with the zoom
A frequent complaint with the latest update is the limited width of the
lectures content on tablets. This was done on purpose both for
aesthetic and ergonomics reasons.

On the aesthetic side, having left aligned text pulled on left with a
significant amount of empty space on the right is not very nice
visually.

On the ergonomic side, when the text is longer but with a default zoom,
the eyes needs to scan and follow a long line which is tiring and
requires an effort.

On the other side, as it turns out, most users buying a tablet do it
precisely to zoom a lot and get a bigger text. To them, having a zoomed
text centered on a smartphone-sized columns is not ideal either.

This commit reconciles both needs by dynamically adjusting the width of
the webview based on the zoom level. Special care has been taken on
handling screen rotation and other width change events.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>