bible v2: add stub Bible v2
We now have Bible section. This update was one of the most praised in
the application history! We want to take it to next level and allow for
early feedback without breaking the main features.

This commit introduces a "clean-sheet" new section to experiment with a
new way to integrate the Bible in the application.

This new section is hidden behind a 'secret' tester option to limit the
impact on existing users.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>