bible v2: display the book list
The display is built around 3 tabs, for the old and new testatement as
well as the psalms. Technically, this is built around a recyclerview.
While not technically mandatory, this is a great occasion to exercise
this component and lear how to use it.

In this first version, we support display section names as well as book
names. This is currently no clic-able and therefore not very useful :)

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>