bible v2: import minimal Bible markup
To share as many code as possible with the existing office display, we
need to re-import and filter the Bible from aelf.org to keep a minimal
markup that is compatible with our existing CSS sheets.

This commit comes with the import code as well as a cached version of
the end result.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>