build: wire up kotlin support
We'll use the next commit to start experimenting with Kotlin on a very
humble module. On the first impression, it was very painful to get
working, with Android Studio pretending that the setup is up and running
when it was not and the cute "1_8" to "18" way of naming java versions
in Android headache.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>