display: drop full-screen mode
The full screen mode was added at the very begining of the developement
of the application. At that time, this was the best and most common way
to reduce distraction when using the application. This mode had the
downside of being un-intuitive to exit, especially for people not used
to games on Android.

Whith newer releases of Android, new innovations painfully arrived bit
by bit and the un-obtrusive "low profile" mode was introduced and
matured. This mode is best suited for reading applications, like this
applications and was therefore quickly adopted.

However, this mode took a couple of versions to settle and the logic to
handle this variety of "full screen" modes became a nighmare to manage.
Maybe we should have done multiple versions of this function with
dynamic dispatch instead of doing all in the same function.

Anyway, 10 years later, the dust has largely settled and the low profile
mode has been mature for years. Time has come to get rid of this mess
and also remove an option from the settings!

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>