lectures: robustify reference extraction
There are occasionally references that are not directly Bible reference.
For example on 2023-09-15 mass' gospel, the reference is prefixed with
"Stabat Mater. ". This defeats the reference parser and causes the link
the break (and even crash, prior to the previous commit). Similarly, on
the following day, the african variant of the gospel verse reference was
"S. Cyprien", which is a valid reference but, obviously, not a valid
/Bible/ reference.

This commit introduces a terrible regexp to extract a reference-ish from
a string suspected to contain a valid Bible reference.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>