reading: disable 'search on web' action
Text selection is enabled in the readings by design. We want to allow
sharing readings. However, the default action menu is not fully
appropriate. Indeed, the 'search on web' action does not mean much in
this context. If a user really wants to do this, the 'share' and 'copy'
actions both remain available as a workaround.

This will free some valuable room for a 'bookmark' option.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>