common: hook into WebView's context menu cycle
A commonly requested feature for the AELF application is a bookmark
feature for the Bible. The envisionned workflow for this feature would
be:

1. A user selects some text
2. The WebView's context menu pops and exposes a "Bookmark" action

This workflow requires to either re-implement Android's stock WebView
context menu to preserve existing features OR hook into the existing
process to inject our custom action.

This commit introduces the necessary plumbing to intercept the context
menu creation and events handling.

Note: The interface dance was needed as the bookmark feature will only
be appropriate to the Bible part.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>