Rework API version handling
There are 3 issues with the current API version handling:

1. The value is propagated via the preferences. This is racy. We have
   cases where the value does not propagate and "-1" is used instead
   which triggers an error on the API server and a crash due to invalid
   Json on the application side.
2. Bumping the API value with each release causes artificial cache
   misses on the servers and increased the load for no added value.
3. The API version is used in too many places

This commit de-couples the API version from the app version and starts
with the last officially released version. It also takes the occasion to
load te version from a single location in the code and then propagate it
as an argument. That's at least one clean-ish thing the in the sync
code...

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>