Refactor the cache around a single table
During the sync, the engine queries the cache for each candidate office.
This takes a very long time merely to see that the readings are here. A
better idea would be to dump the list of cached offices from the cache
in a single request. A first step toward this goal is to move to a
single table instead of one table per office.

As a nice side-effect, this allowed to simplify the code.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>