Speed up cache refresh by 40s
On typical cache refresh, at most the current week and an additional day
needs to be fetcjhed from the network. Up to now, the check to decide if
we should fetch from the network was implemented as an office load from
the cache. This is extremely costly. It requires compiling a SQL
statement, executing it and even deserializing the office, for each
candidate.

Instead, we only need to fetch the list of cached antries once and fetch
the missing entries. This is what this commit does.

The ultimate goal is to reduce power usage to reduce the risk of being
limited by Android and speed up the sync to limit the risk of moving
away from a reliable connection during the sync.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>