Store and retrieve lectures checksums in the cache
The server side provides office checksums in the form of "Etag". This
checksum is reproducible on server side and only changes when the
content of the office changes. That is, when the office was updated.

The sync engine loads all offices for the month ahead (by default) and
unconditionally re-loads the offices of the week to come. This puts a
large load on the server and also keeps the phone awake longer.

The idea is to use these checksums to determine if an office is up to
date or needs to be refreshed.

This commit wires the full office checksum lifecycle in the application.
The checksums are saved and loaded to/from the cache.

A future commit will wire a call to the API to get the list of the
office as currently available on the server side. This will allow the
sync engine to only reload changed offices.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>