Load office checksums from the API
As a major step toward optimizing the sync of the 7 next days, this
commit wires the full logic to load and de-serialize the checksums from
the API.

As this was pretty challenging to get right, this was the occasion to
(at long last) introduce the first unit tests. This was not trivial
because the code base is 10 years old and is not that well
architectured. At least, that's a significant step in the right
direction!

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>