Only refresh next 7 days if the server reports them as outdated
This commit is culminating point of the previous series of commit. With
this work, a no-op refresh is now closer to 2-3s on my phone than the
previous 20-25s when the next seven days were always refreshed and even
further from the 1m20s before we started this optimization work.

With this commit, the next 7 days are always checked for freshness
according to the current server status but only refreshed in case of
mismatch.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>