Implement the new sync engine
The previous sync engine was based on the once-standard sync adapters
running in a dedicated process.

This caused a couple of drawbacks:
- We had all challenges related to inter-process
  - Shared SQlite instance
  - Painful debugging
  - ...
- On some devices the sync does not work at all, for unknown reason

On top of that, the code structure did not move much over the past 9
years and was clearly showing signs of age.

This new commit splits the sync manager from the sync worker itself and
defers network management to the WorkManager. This clarifies
responsibilities and removes clutter. Additionally, the code was largely
refactored to better highlight te units of work. Last but not least, the
sync engine coordination has been moved from the activity to the
application, where it belongs.

Hopefully, we should be ready for another 10 years ;)

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>