Improve cache robustness
The cache main class is one of the first piece of code of this
application. It was ill-named and the database upgrade/delete path was
not properly integrated with the base class.

This commit renames the class to be more consistent with its role and
the conventions in the application. It also now relies on the base class
to delete the database on upgrade, rather than doing it ourselves. Same
goes for error recovery.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>