Track office generation time to break ties on checksum mismatch
When the server and the local cache disagree on the office checksum,
this may be caused by two possible reasons:

* Either the office is really outdated
* Or the office checksums cache is stale on the server side (up to
  4h in the worse case with today's cache configuration)

In the second case, if we refresh the cache, we'll end up with another
cache mismatch. To break this tie, the server has added a
"generation-date" field in the cache manifest API and a "Last-Modified"
HTTP header on API responses. The header is copied in the local cache
and then compared to the date provided by the API.

And this was extremely painful to implement in Java, especially with the
date mess. But the newly introduced tests made it tractable.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>