Insert complete 2-verses variant of the Doxology
There are 3 variants of the Doxology, a 2, 3 and 4 verses one. The
selection depends on the chosen psalm rythms. Up to now we were only
inserting the beginning to suggest it.

However, the main aim of this application is to help occasional hours
priar. Since the 2-verses version is the most common, we insert this one
and let more experts switch to the appropriate variant.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>