Fix sticky title when comming back from the Bible
Previously, when comming back from the Bible, the office position was
lost. While fixing this bug, another one was introduced (clever design)
that caused the title of the Bible book to remain and not being replaced
by the office title.

The root cause of this is the mixup between data, UI and state
management. But this commit only applies duct tape. We already worked
quite a bit on the infra for now.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>