Enable predictive back
Android 15 will introduce predictive back, that is, a pre-visualization
of the destination of the back button. This is pretty nice to avoid
accidentally exiting the application.

Whiel Android 15 is nowhere near release, this change will be highly
encouraged and this is the occasion to clean up the essentially
deprecated calls used in the back stack management.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>