Prototype drawer header - static & light only
The "Information" section has always been a hack. Since the very start
of the application, the idea was to, somehow, integrate it visually.

A good place for this is the drawer header which is currently unused.
The drawer header is prototyped to highlight the most important
information to locate the current day in the Missel and Psautier. That
is:

* The day
* The liturgical years and Psautier week number
* Each possible propers and associated liturgical color

This commit introduces a fully hard-coded initial version for dicussion.
Future commits will bring proper theming, data integration and Error fallback.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>