Fix truncated bible book text name on overflow
When the device font size has been increased at the device level, the
Bible book names may overflow. Currently, the height of the book items
is fixed and terefore, when the book name wraps, the second line is only
partially visible.

This commit moves to a fixed /minimum/ height and otherwise moves to
"wrap_content".

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>