Introduce new data model for informations
The information data model has been overhauled on the server side to
support the drawer header feature. This commit introduces the
corresponding suppoirt on the client side and bumps the database version
to trigger a full invalidation. Whithout the invalidation, there would
be an exception on loading from the cache due to the class change.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>