Automatically fill the office drawer header
Now that we have the layouts, automatic header selection and the
datasource, this commit puts it all together and fills the header with
the daily information.

Known issue: When selecting a date using the calendar widget, the update
is only visible after the office is loaded. In case of a load from
network, this may imply a visual delay.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>