Refactor WebView lifecycles
We have a lot of ANRs related to the webview. Looking more deeply
around, the application starts with 4 webviews and can use up to 6 in
the worst case in the Bible. These 4 webviews are typically instanciated
at the same time, possibly concurrently with the initial sync.

Every time a page is swiped, a date changed, an office selected, ... new
webviews are instanciated on the fly, while only a pool of up to 6 is
needed in practice.

The plan is to switch to a pool of webviews. To support this, we first
need to dynamically attach / detach the webview and fully encapsulate
its lifecycle.

This commit does just that.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>