Reduce sync threads priority
There is more jank when the sync is running in the background. It is
important to run the sync in the background, when the application is
exposed to the user. This is important, because low end phones tend to
aggressively kill the application when going in the background thus
defeating the purpose of background sync.

This commit attempts to partially mitigate the situation by lower the
priority of the threads.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>