Fix WebView leak
While investigating the WebView ANRs reports, I realized that we have a
WebView leak (actually, full fragments) when going back. The forward
path is fine.

In the back path, new views and pager adapter are created. While
allocating a new adapter and attaching it to the pager, like is done in
the "refresh" path, does not trigger the bug, doing it on a fresh
instance of the pager does trigger the leak.

The correct fix is to detach the pager adapter from the pager when the
view is destroyed.

With this fix, the absolute peak number of webviews is now down to 6, in
the Bible, when jumping from a middle chapter to another middle chapter.
In then settles back to 3, which is the general case.

A good reproducer was to move back and forth: navigate to a new date,
office or section and then back. Repeating this process would just
increase the leak. I stopped at 75 instances which doubled the memory
requirements and made the "back" awfully slow.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>