Scale sync workers count with hardware
Since the last release, we get a lot of ANR in the webview render phase.
Not much changed in the webview related logic in the past release,
however a significant change was the launch of the sync in the forground
activity on initial start and when the OS failed to schedule the sync
(low-end phones). Additionally, as part of the effort to improve the
sync performance, the sync was moved from single thread to 4 threads,
based on my high-end phone experience.

In the absence of reliable data, a reasonable guess would be that this
initial sync and multi-threading is possible cause for the ANR as it
puts more stress on all system resources (CPU and memory at least).

This commit attempts to guess the device performance class based on
logic borrowed from the Telegram application and uses this as the basis
to scale the number of workers.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>