Introduce a WebView pool
The application is build around "slides". Each slide is basically
composed of a WebView for rich text rendering.

For performance and memory management reasons, we only allocate what is
needed for a smooth render. In the general case, we need 3 slides. The
visible one and the 2 "side" slides. When sliding a transient 4th one
is generated. If slideing to a random location, like a jump of chapter
in the Bible, 3 more needs to be allocated, thus getting to a total of
6 in the worse case scenario.

In the current system, everytime we need a webview, a fresh one is
allocated thus putting pressure on the initialization and GC.

This commit introduces a webview pool. The pool is initialized to fit
the general case when sliding in the middle of an office. That is, 4
webviews. It supports allocating up to 6 cacheable webviews and then
falls back to on-demand throw-away allocation to avoid crashing if the
pool was not correctly sized.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>