Delay background sync
The initial sync is the most likely cause of the very large number of
ANRs. The previous attempts to optimize did not help much.

This coommit attempts a different approach. It introduces a 5 minute
initial delay to avoid colliding with intial application open while
remaining in the foreground time window.

Additionally, the sync parallellism has been reduced as an effort to
reduce the load even more.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>