Fix crash when rotating the screen
The move to a webview pool introduced a new crash when rotating the
screen after a webview has been put back in the pool.

In this case, for an unexplained reason, the system calls `getDisplay`
on the webview context, even though it is detached from the hierarchy.
This triggers an exception.

This commit introduces a stub display based on a virtual surface. When
the exception is detected, the surface is returned instead.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>