Always center tabs
Years ago, when there was a single tab, the tab would be centered. This
was lost when moving to androidx. It turns out the response was right in
front of my eyes in the logs lines:

```
MODE_SCROLLABLE + GRAVITY_FILL is not supported, GRAVITY_START will be used instead
```

This commit cleans the logs and centers tabs when there is only one.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>