Move back to async lecture loading in all cases
We have a lot of ANR since v2.8.5. The first suspect was the new sync
engine. However, on second check, the engine is not running when the ANR
is detected.

Another option could be the accidental move to synchronous lecture view
pager setup in the back path.

This change was made to preserve the current slide when navigating from
a back button. It also introduced terrible additional complexity in an
already overly complex historical class full of corner cases.

This commit reverts to the async path, to the cost of additional
flicker. However, the position is not preserved in back navigation AND
refresh (which should always have been the case).

This change is made by considering the "whatwhen" part of the state.
This is possible because a frangment instance can only target a single
date + office (Historically, when the code was fully in the main
activity, this was not the case, and we had to deal with change in the
date in the object).

A future refactoring should move the position, cache and other related
settings fully out of "whatwhen" where they never belonged...

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>