Drop useCache and position from WhatWhen
WhatWhen is meant to describe a location in an office. Back when the
application was fully contained in a single activity, it was piggy
backed to track additional UI and intent settings.

"useCache" was a one time toggle that was immediately reset. In other
words, it should always have been... a parameter.

"position" was the UI location index. It is used as a backup of the
current position when saving the state as well as a translation of the
"anchor" field into a UI state. It has been moved to a private field in
the class.

With this work, the class is back to its original vocation there is less
modification to shared state.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>