Support external links in offices
The RSS API is being deprecated and we have no way to provide users with
a simple "1-click" button to trigger to upgrade because all links are
disabled.

While there is no way to fix a bug in a version in the field, let's
enable the feature for future versions.

With this commit, any link will open using the default Android handler
by default.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>