Fix opening of AELF website when trying to see text in context
When clicking on a reference from lecture, the application is supposed
to open the text in its context. With the addition of the feature to
open an arbitrary link, a second intent is generated. This triggers the
opening of the browser in addition to the in-app navigation.

This commit fixes the condition so that only the internal navigation
occurs.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>