Add liturgical color tooltip
A very valuable comment from a beta tester made me realize that the new
information screen is not color blind friendly. Indeed, the liturgical
color is displayed as a colored square while it was mentionned by name
before that.

As an accessibility tweak, this commit introduces a tooltip with the
liturgical color name. This tooltip is exposed using standard Android
techniques and should therefore be reasonably easy to discover.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>