Add dark mode splashscreen
When the application starts, it displays a "white flash" on dark mode
devices as the splash screen is displayed.

The challenge is that, in such an early phase, we do not know yet which
theme the user selected. However, the intresting thing is that we always
override the theme as soon as our code starts to execute.

The trick to set a dark splash screen is then to pretend that our
application main theme is the splash screen, and have it follow the
system theme.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>