fix missing buttons in light theme
The dialog buttons disappeared in the light theme as a regression of the
theme refactoring for the splash screen. This is fixed by making the
splash screen inherit from the main themes. Surprisingly, it does not
break when the phone is in dark mode and the application in light mode
and vice-versa.

The window colors are also tweaked. This makes the drawer "white" less
aggressive in light mode.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>