feat: enhance Bible book titles consistency
A contributor raised a very valid point on the iOS flavor of the
application here https://gitlab.com/nathanael2/aelf-flutter/-/merge_requests/47.

Indeed, the AELF Website, the unofficial applications and the printed
version published by Mame all use different wording for the book names.

The applications are meants to be an offline version of the Website. As
such, application content should be as close as reasonably possible to
the content provided on the website.

As for the name of the books of the Old Testatment, this should be fine.
However, regarding the New Testament, the names chosen on the website
are terribly long and painful to display on a small screen, even in the
mobile version of the website.

For the old testament, this commit applies the website wording. For the
new one, it proposes a "factorization", using the intermediate titles.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>