fix: handle the notch in landscape mode
Starting with Android 15, the application draws edge-to_edge by default.
This is mostly fine but causes an issue with the text in landscape mode,
where a part might be masked by the notch.

This commit overhauls this part. The code for handling tablet and phones
is now fully unified, using a constraint layout and a window inset
listener.

Signed-off-by: Jean-Tiare Le Bigot <jt@yadutaf.fr>